Extract SimpleSchemaLoader to its own file
